+++ src/components/pages/select/SettingCard.tsx
